refactor: 팩토리 패턴 수정
- 클래스 생성자의 접근 제어자를 private로 설정하고, 정적 팩토리 패턴을 통해 생성할 수 있도록 합니다.
