+++ src/main/java/lotto/controller/BonusController.java
@@ -1,7 +1,6 @@
 package lotto.controller;
 
 import lotto.domain.Lotto;
-import lotto.factory.BonusGeneratorFactory;
 import lotto.message.PrintMessage;
 import lotto.service.generator.BonusGenerator;
 import lotto.view.InputView;
@@ -31,7 +30,7 @@ public class BonusController {
         try {
             outputView.printlnMessage(PrintMessage.INPUT_LOTTO_BONUS_NUMBER);
             String lottoBonus = inputView.inputUser();
-            return BonusGeneratorFactory.create(winning, lottoBonus);
+            return BonusGenerator.create(winning, lottoBonus);
         } catch (IllegalArgumentException e) {
             outputView.printException(e.getMessage());
         }
